+++ src/TerrainMap.h
@@ -14,8 +14,8 @@ class TerrainMap
 
         int coord_to_virtual_bitmap_x(int x, int y);
         int coord_to_virtual_bitmap_y(int x, int y);
-        double virtual_bitmap_to_coord_x(int x, int y);
-        double virtual_bitmap_to_coord_y(int x, int y);
+        int virtual_bitmap_to_coord_x(int x, int y);
+        int virtual_bitmap_to_coord_y(int x, int y);
 
     protected:
     private:
